Move: LabelBottomNaviation.tsx의 디렉터리 변경
아직 다른 컴포넌트들도 디렉터리 변경이 필요하지만 일단 내가 추가한 기능과 연괸있는 컴포넌트만 이동함

> LabelBottomNavigation.tsx에서 onClick()은 redux slice에서 change함수를 통해 상태를 변경하게 했음

다음 커밋과 이어짐
