+++ WeMeet/src/main/wemeet_front/src/features/bottomNav/LabelBottomNavigation.tsx
@@ -3,8 +3,11 @@ import {BottomNavigation, BottomNavigationAction} from "@mui/material";
 import AssistantDirectionIcon from "@mui/icons-material/AssistantDirection";
 import MapIcon from "@mui/icons-material/Map";
 import ExploreIcon from "@mui/icons-material/Explore";
+import {useDispatch} from "react-redux";
+import {change} from "./bottomNavSlice";
 
 export default function LabelBottomNavigation() {
+    const dispatch = useDispatch();
     const [value, setValue] = useState("map");
 
     const handleChange = (event: React.SyntheticEvent, newValue: string) => {
@@ -15,9 +18,15 @@ export default function LabelBottomNavigation() {
                 value={value}
                 onChange={handleChange}
             >
-                <BottomNavigationAction label="코스 추천" value="courseSugg" icon={<AssistantDirectionIcon />} />
-                <BottomNavigationAction label="지도" value="map" icon={<MapIcon />} />
-                <BottomNavigationAction label="피플픽" value="peoplePick" icon={<ExploreIcon />} />
+                <BottomNavigationAction label="코스 추천" value="courseSugg" onClick={()=>{
+                    dispatch(change(0))
+                }} icon={<AssistantDirectionIcon />} />
+                <BottomNavigationAction label="지도" value="map" onClick={()=>{
+                    dispatch(change(1))
+                }} icon={<MapIcon />} />
+                <BottomNavigationAction label="피플픽" value="peoplePick" onClick={()=>{
+                    dispatch(change(2))
+                }} icon={<ExploreIcon />} />
             </BottomNavigation>
     );
 }
